+++ src/MailGlyph.php
@@ -9,6 +9,7 @@ use MailGlyph\Resources\Contacts;
 use MailGlyph\Resources\Emails;
 use MailGlyph\Resources\Events;
 use MailGlyph\Resources\Segments;
+use MailGlyph\Resources\Templates;
 
 final class MailGlyph
 {
@@ -19,6 +20,8 @@ final class MailGlyph
     public readonly Contacts $contacts;
 
     public readonly Campaigns $campaigns;
+
+    public readonly Templates $templates;
 
     public readonly Segments $segments;
 
@@ -36,6 +39,7 @@ final class MailGlyph
         $this->events = new Events($httpClient);
         $this->contacts = new Contacts($httpClient);
         $this->campaigns = new Campaigns($httpClient);
+        $this->templates = new Templates($httpClient);
         $this->segments = new Segments($httpClient);
     }
 }